CoreFeatures, add action_index() to delegate to its action_features().
This is simply a level of indirection, when the work is all done in
action_features; core features is a form which accepts user input
to switch on/off. If/when we split the user input handling from the
rest of action_features(), we might have a benefit. Now it's just
a matter of standardization attempt: action_index() directs to stuff,
doesn't process usually.

Signed-off-by: Norv <a.w.norv@gmail.com>