Guest Posting: removed colons, aligned input box
Added the space in the string ( sorry emanuele ) so the Email input box
would line up with the Name input box. Changed one string to fix its
meaning. ( sorry again emanuele ). Added 3 strings to be used in
following commits.  It seems the name and email strings are used a lot
and can cause issues with and without the colon.