Fonts conflicting with non inherited fonts
These two deletions were causing the input boxes for the guest
name/email on the verification section to be off centered from each
other. This only occurs in the besocial variant as fonts are invoked or
inherited in about 5 places.  As well, removing these two lines help to
middle align some of the preset options for the drop down boxes.